Gracefully handle security settings preventing local/session storage
Closes #431
