+++ ui/frontend/storage.spec.ts
@@ -1,18 +1,6 @@
 import { createStore } from 'redux';
 
-import storage from './storage';
-
-class TestStorage {
-  private data = {};
-
-  public getItem(name: string): string {
-    return this.data[name];
-  }
-
-  public setItem(name: string, value: string) {
-    this.data[name] = value;
-  }
-}
+import storage, { InMemoryStorage } from './storage';
 
 const identityReducer = <S>() => (s: S, _a: any): S => s;
 const serialize = JSON.stringify;
@@ -20,14 +8,14 @@ const deserialize = JSON.parse;
 
 describe('restoring saved state', () => {
   test('partially serialized data is merged with initial state', () => {
-    const testStorage = new TestStorage();
+    const testStorage = new InMemoryStorage();
     testStorage.setItem('redux', serialize({ config: { alpha: true } }));
 
     const initialState = {
       config: { alpha: false, beta: 42 },
     };
 
-    const enhancer = storage({ storage: testStorage, serialize, deserialize });
+    const enhancer = storage({ storageFactory: () => testStorage, serialize, deserialize });
     const store = createStore(identityReducer<typeof initialState>(), initialState, enhancer);
 
     const state = store.getState();
