+++ ui/frontend/storage.ts
@@ -4,9 +4,21 @@ import { StoreEnhancer, StoreEnhancerStoreCreator } from 'redux';
 type SimpleStorage = Pick<Storage, 'getItem' | 'setItem'>;
 
 interface Config<S> {
-  storage: SimpleStorage;
+  storageFactory: () => SimpleStorage;
   serialize: (state: S) => string;
   deserialize: (state: string) => S;
+}
+
+export class InMemoryStorage {
+  private data = {};
+
+  public getItem(name: string): string {
+    return this.data[name];
+  }
+
+  public setItem(name: string, value: string) {
+    this.data[name] = value;
+  }
 }
 
 const key = 'redux';
@@ -14,16 +26,22 @@ const key = 'redux';
 const storage = <St>(config: Config<St>): StoreEnhancer =>
   (createStore: StoreEnhancerStoreCreator<{}, St>) =>
     (reducer, preloadedState) => {
-      const { storage, serialize, deserialize } = config;
+      const { storageFactory, serialize, deserialize } = config;
 
-      let serializedState;
+      let storage: SimpleStorage;
 
       try {
-        serializedState = storage.getItem(key);
+        // Attempt to use the storage to see if security settings are preventing it.
+        storage = storageFactory();
+        const current = storage.getItem(key);
+        storage.setItem(key, current);
       } catch (e) {
-        console.warn(`Unable to load initial state from ${storage}: ${e}`); // tslint:disable-line:no-console
+        // tslint:disable-next-line:no-console
+        console.warn('Unable to store configuration, falling back to non-persistent in-memory storage');
+        storage = new InMemoryStorage();
       }
 
+      const serializedState = storage.getItem(key);
       const persistedState = deserialize(serializedState);
       const mergedPreloadedState = merge(preloadedState, persistedState);
       const theStore = createStore(reducer, mergedPreloadedState);
@@ -31,12 +49,7 @@ const storage = <St>(config: Config<St>): StoreEnhancer =>
       theStore.subscribe(() => {
         const state = theStore.getState();
         const serializedState = serialize(state);
-
-        try {
-          storage.setItem(key, serializedState);
-        } catch (e) {
-          console.warn(`Unable to save state to ${storage}: ${e}`); // tslint:disable-line:no-console
-        }
+        storage.setItem(key, serializedState);
       });
 
       return theStore;
